requirements.txt
flask==3.0.2
itsdangerous==2.1.2  # Required for session management
requests==2.31.0  # For API calls
python-dotenv==1.0.1  # For environment variables 